requirements.txt
openai>=1.0.0
gspread>=5.0.0
oauth2client>=4.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
langgraph>=0.2.0
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
numpy>=1.24.0